move get_experimet_type from data to data creator quantstudio design
