+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/appbio_quantstudio_designandanalysis_structure.py
@@ -578,35 +578,3 @@ class Data:
     calculated_documents: list[CalculatedDocument]
     reference_target: str | None
     reference_sample: str | None
-
-    @staticmethod
-    def get_experiment_type(contents: DesignQuantstudioContents) -> ExperimentType:
-        experiment_type_to_expected_sheets = {
-            ExperimentType.standard_curve_qPCR_experiment: ["Standard Curve Result"],
-            ExperimentType.relative_standard_curve_qPCR_experiment: [
-                "RQ Replicate Group Result"
-            ],
-            ExperimentType.genotyping_qPCR_experiment: ["Genotyping Result"],
-            ExperimentType.melt_curve_qPCR_experiment: [
-                "Melt Curve Raw",
-                "Melt Curve Result",
-            ],
-            ExperimentType.presence_absence_qPCR_experiment: [
-                "Sample Call",
-                "Well Call",
-                "Target Call",
-                "Control Status",
-            ],
-        }
-
-        possible_experiment_types = {
-            experiment_type
-            for experiment_type, expected_sheets in experiment_type_to_expected_sheets.items()
-            if all(contents.has_sheet(sheet_name) for sheet_name in expected_sheets)
-        }
-
-        if len(possible_experiment_types) == 1:
-            return possible_experiment_types.pop()
-
-        msg = f"Unable to infer experiment type from sheets in the input, expected exactly one set of sheets from: {list(experiment_type_to_expected_sheets.values())}, got {list(contents.data.keys())}"
-        raise AllotropeConversionError(msg)
